+++ sling/nlp/parser/trainer/cascade.py
@@ -388,11 +388,12 @@ class Cascade(object):
     return s
 
   """Saves the cascade specificaton to a frame in 'store'."""
-  def as_frame(self, store):
+  def as_frame(self, store, delegate_cell_prefix):
     frame = store.frame({"id": "cascade", "name": self.__class__.__name__})
     delegates = store.array(self.size())
     for index, delegate in enumerate(self.delegates):
       d = store.frame({"name": delegate.__class__.__name__, "index": index})
+      d["cell"] = delegate_cell_prefix + str(index)
       delegate.as_frame(d)
       delegates[index] = d
     frame["delegates"] = delegates
